more config parameters, fixed #147
- config for spear
- config for mining speed
- config for bonus attack damage (currently only modifies spear)

+++ common/src/main/java/net/enderitemc/enderitemod/tools/EnderiteSword.java
@@ -28,7 +28,7 @@ import org.jspecify.annotations.Nullable;
 
 public class EnderiteSword extends Item {
 
-    public EnderiteSword(ToolMaterial material, int attackDamage, float attackSpeed, Item.Settings settings) {
+    public EnderiteSword(ToolMaterial material, float attackDamage, float attackSpeed, Item.Settings settings) {
         super(settings.sword(material, attackDamage, attackSpeed));
     }
